Add files via upload
The challenge was to build out this card component and get it looking as close to the design as possible, use any tools you like to help you complete the challenge. So if you've got something you'd like to practice, feel free to give it a go.
Users should be able to:
- View the optimal layout depending on their device's screen size (375px & 1440px). 
Feedback is more than appreciated.